backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
flask-socketio==5.3.6
gevent==24.2.1
gevent-websocket==0.10.1
sqlalchemy==2.0.25
mysql-connector-python>=9.0.0
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
apscheduler==3.10.4
numpy==1.26.4
cryptography>=41.0.0
